Renaming Registration folder to Portal_helper
