+++ lib/testing_site_onlyoffice/portal_helper/portal_helper.rb
