+++ lib/testing_site_onlyoffice/portal_helper/portal_version.rb
